feat: isActive checks, accept/reject task controllers added

+++ controllers/client/controllers.js
@@ -5,10 +5,11 @@ const Task = require("../../models/tasks");
 // route: GET /api/client/getAllTask
 exports.getAllTask = async (req, res) => {
   const { email } = req.user;
+  const { status = "created" } = req.query;
 
   try {
     const tasks = await Task.aggregate([
-      { $match: { email } },
+      { $match: { email, status } },
       { $addFields: { images: { $size: "$images" } } },
       { $sort: { createdAt: -1 } },
       { $project: { _id: 0, email: 0, __v: 0 } },
@@ -82,8 +83,17 @@ exports.deleteTask = async (req, res) => {
   const { taskId } = req.params;
 
   try {
-    await Task.deleteOne({ id: taskId });
-    return res.status(200).json({ message: "Task deleted successfully" });
+    const task = await Task.findOne({ id: taskId });
+    if (!task) {
+      return res.status(404).json({ message: "Task not found" });
+    }
+
+    // only inactive task, can be deleted
+    if (!task.isActive) {
+      await Task.deleteOne({ id: taskId });
+      return res.status(200).json({ message: "Task deleted successfully" });
+    }
+    return res.status(400).json({ message: "Task cannot be deleted" });
   } catch (error) {
     console.error("Error deleting task:", error);
     return res.status(500).json({ message: "Failed to delete task" });
@@ -121,7 +131,9 @@ exports.updateTask = async (req, res) => {
         return res.status(400).json({ message: "No fields to update" });
       }
     }
-    return res.status(400).json({ message: "Failed to update.Task is active" });
+    return res
+      .status(400)
+      .json({ message: "Failed to update. Task is active" });
   } catch (err) {
     console.error("Error updating task:", err);
     return res.status(500).json({ message: "Failed to update task" });
